--- MobCore/src/main/java/eu/cloudnetservice/cloudmob/MobConfig.java
@@ -1,42 +1,0 @@
-package eu.cloudnetservice.cloudmob;
-
-import com.google.gson.reflect.TypeToken;
-
-import java.lang.reflect.Type;
-import java.util.Map;
-
-public class MobConfig {
-
-    public static final Type TYPE = TypeToken.get(MobConfig.class).getType();
-    private int inventorySize;
-
-    private int startPoint;
-
-    private MobItemLayout itemLayout;
-
-    private Map<Integer, MobItemLayout> defaultItemInventory;
-
-    public MobConfig(int inventorySize, int startPoint, MobItemLayout itemLayout, Map<Integer, MobItemLayout> defaultItemInventory) {
-        this.inventorySize = inventorySize;
-        this.startPoint = startPoint;
-        this.itemLayout = itemLayout;
-        this.defaultItemInventory = defaultItemInventory;
-    }
-
-    public int getInventorySize() {
-        return inventorySize;
-    }
-
-    public int getStartPoint() {
-        return startPoint;
-    }
-
-    public Map<Integer, MobItemLayout> getDefaultItemInventory() {
-        return defaultItemInventory;
-    }
-
-    public MobItemLayout getItemLayout() {
-        return itemLayout;
-    }
-
-}
