+++ MobModule/src/main/java/eu/cloudservice/v2/mobs/in/PacketInRemoveMob.java
@@ -1,0 +1,15 @@
+package eu.cloudservice.v2.mobs.in;
+
+import de.dytanic.cloudnet.lib.network.protocol.packet.PacketInHandler;
+import de.dytanic.cloudnet.lib.network.protocol.packet.PacketSender;
+import de.dytanic.cloudnet.lib.utility.document.Document;
+import eu.cloudnetservice.v2.mobs.ServerMob;
+import eu.cloudservice.v2.mobs.MobModule;
+
+public class PacketInRemoveMob extends PacketInHandler {
+
+    public void handleInput(Document packet, PacketSender packetSender) {
+        ServerMob serverMob = packet.getObject("mob", ServerMob.TYPE);
+        MobModule.getInstance().getMobDatabase().remove(serverMob);
+    }
+}
