--- MobPlugin/src/main/java/eu/cloudservice/cloudmob/event/BukkitMobUpdateEvent.java
@@ -1,29 +1,0 @@
-package eu.cloudservice.cloudmob.event;
-
-import de.dytanic.cloudnet.bridge.event.bukkit.BukkitCloudEvent;
-import eu.cloudnetservice.cloudmob.ServerMob;
-import org.bukkit.event.HandlerList;
-
-public final class BukkitMobUpdateEvent extends BukkitCloudEvent {
-
-    private static HandlerList handlerList = new HandlerList();
-
-    private ServerMob serverMob;
-
-    public BukkitMobUpdateEvent(ServerMob serverMob) {
-        this.serverMob = serverMob;
-    }
-
-    public static HandlerList getHandlerList() {
-        return handlerList;
-    }
-
-    public ServerMob getServerMob() {
-        return serverMob;
-    }
-
-    @Override
-    public HandlerList getHandlers() {
-        return handlerList;
-    }
-}
